deployment/pyproject.toml
[project]
name = "deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.121.0",
    "openai-agents>=0.5.0",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
